JWT token authorization added to endpoints

+++ src/BLL/CoffeeRoastery.BLL.Interface/Dto/JwtTokenResponse.cs
@@ -3,6 +3,6 @@ namespace CoffeeRoastery.BLL.Interface.Dto;
 public class JwtTokenResponse
 {
     public string AccessToken { get; set; }
-    public int ExpiresIn { get; set; }
+    public double ExpiresIn { get; set; }
     public string TokenType { get; set; }
 }
